working ocean current model plugin
